requirements.txt
flask==2.0.1
langchain==0.0.184
langchain-community==0.0.6
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
python-docx==0.8.11
chardet==4.0.0
werkzeug==2.0.1
chromadb==0.3.21
python-dotenv==0.19.0
pytest==6.2.5